Do not wait needlessly on phantomjs
Seems like phantomjs exits when done, so it's better to just poll every once
in a while the process, and if it's done, to just continue, and only kill it
if it sleeps for more than 90 seconds. This helps noticeably and also doesn't
leave a ton of children in zombie state.